typehead and datepicker - common component created and integrated with new expense
